requirements.txt
aiohttp==3.9.5
prometheus-client==0.20.0
PyYAML==6.0.2
databricks-zerobus-ingest-sdk==0.2.0
databricks-sdk>=0.20.0
asyncua==1.1.5
aiomqtt==2.0.1
pymodbus==3.6.4
pysparkplug==0.6.1
